+++ README.md
@@ -15,15 +15,9 @@ mongod -dbpath <<absolute path to your data directory>>
   * create the following document:
   ```
   image = {
-    "filename" : "photo-1581092930041.png", 
+    "filename" : "SSDLC_Pedal_View.png", 
     "photo_name" : "Security Touchpoints in Pedal",
     "album" : "tech drawings", 
-    "description": "Pedal is an Agile, student oriented, software development model. This diagram illustrates McGraw's security touchpoints mapped on to Pedal.",
-    "f_stop": "",
-    "s_speed: "",
-    "focus_len": "",
-    "iso": "",
-    "camera": "",
     "upload_date" : ISODate()
   }
   ```
@@ -31,7 +25,7 @@ mongod -dbpath <<absolute path to your data directory>>
 5. Open a terminal (command shell) and change directory to the working directory of your project.
 6. Start the web server with either the command `npm start` or `node start`.
 7. To run the app
-  * point your browser http://localhost:3000/upload.html,
+  * point your browser http://localhost:3000,
   * choose an image and complete the form, and
   * upload the image.
 8. ~~(**Current Version**) Pointing your browser to http://localhost:3000/getImages will return a JSON array of images you have saved in the database.~~
