--- configuration/config.json
@@ -1,3 +1,0 @@
-{
-    "secret": "5JgKvsOxd278oarolsZ45Me9LkjEifmH"
-}
