+++ package.json
@@ -10,10 +10,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.19.0",
     "express": "^4.16.3",
-    "jwt-simple": "^0.5.6",
     "mongoose": "^5.7.13",
     "multer": "^1.3.0",
     "path": "^0.12.7"
