--- public/css/upload.css
@@ -1,3 +1,0 @@
-body {
-    background-color: darkturquoise;
-}
