--- public/upload.html
@@ -1,71 +1,0 @@
-<!--suppress ALL -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Upload Images to Server</title>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css"
-          integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
-
-    <link rel="stylesheet" href="css/upload.css">
-    <!-- <style>
-        body {
-            background: darkturquoise;
-        }
-    </style> -->
-</head>
-<body>
-
-<div class="container">
-    <div class="card border-success mt-5">
-        <h1 class="card-header">Upload Image</h1>
-        <div class="card-body">
-            <div class="form-group">
-                <form action="api/images" method="post" enctype="multipart/form-data">
-                    <label for="photo">Image File</label>
-                    <input class="form-control-file" type="file" accept="image/*" name="photo" >
-                    <label for="photoName">Image Name </label>
-                    <input class="form-control" type="text" name="photoName" id="photoName" >
-                    <label for="subject">Image Album </label>
-                    <input class="form-control" type="text" name="album" id="album" >
-                    <label for="description">Image Description </label>
-                    <textarea class="form-control" rows="5" name="description" id="description" ></textarea>
-                    <label for="dateTaken">Date Taken or Created</label>
-                    <input class="form-control" type="date" name="dateTaken" id="dateTaken" >
-                    <fieldset class="form-control">
-                        <legend>Photo Parameters (optional)</legend>
-                        <div class="form-row">
-                            <div class="form-group col-md-3">
-                                <label for="f_stop">F-stop</label>
-                                <input class="form-control" type="text" name="f_stop" id="f_stop" >
-                            </div>
-                            <div class="form-group col-md-3">
-                                <label for="s_speed">Shutter Speed</label>
-                                <input class="form-control" type="text" name="s_speed" id="s_speed" >
-                            </div>
-                            <div class="form-group col-md-3">
-                                <label for="iso">ISO</label>
-                                <input class="form-control" type="text" name="iso" id="iso" >
-                            </div>
-                            <div class="form-group col-md-3">
-                                <label for="focus_len">Focus Length</label>
-                                <input class="form-control" type="text" name="focus_len" id="focus_len" >
-                            </div>
-                        </div>
-                        <label for="camera">Camera</label>
-                        <input class="form-control" type="text" name="camera" id="camera" >
-                    </fieldset>
-                    <input class="btn-default mt-3" type="submit" value="Save">
-                    <input class="btn-default mt-3" type="reset" value="Clear">
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
-
-</body>
-</html>
